move `CompilationOptions` to `widgets/` and use a template for the button
