+++ static/panes/compiler.ts
@@ -61,12 +61,12 @@ import {toolsService} from '../services/tools.service.js';
 import {SiteSettings} from '../settings.js';
 import * as utils from '../utils.js';
 import {Alert} from '../widgets/alert.js';
+import {CompilationOptions} from '../widgets/compilation-options.js';
 import {CompilerPicker} from '../widgets/compiler-picker.js';
 import {WidgetState} from '../widgets/libs-widget.interfaces.js';
 import {LibsWidget} from '../widgets/libs-widget.js';
 import * as TimingWidget from '../widgets/timing-info-widget.js';
 import {Toggles} from '../widgets/toggles.js';
-import {CompilationOptions} from './compilation-options.js';
 import {CompilerCurrentState, CompilerState} from './compiler.interfaces.js';
 import {GccDumpFiltersState, GccDumpViewSelectedPass} from './gccdump-view.interfaces.js';
 import {LeanCOptions} from './leanc-view.interfaces.js';
