+++ views/templates/panes/compiler.pug
@@ -11,8 +11,7 @@ mixin newPaneButton(classId, text, title, icon)
         button.btn.btn-sm.btn-light.input-group-text.picker-popout-button(data-trigger="click" style="cursor: pointer;" role="button" title="Compiler picker popout")
           span
             i.fa-solid.fa-arrow-up-right-from-square
-        button.btn.btn-sm.btn-light.input-group-text.prepend-options(data-trigger="click" style="cursor: pointer;" role="button" title="All compilation options")
-          span.status-icon
+        include ../widgets/compilation-options
         input.options.form-control(type="text" placeholder="Compiler options..." size="256" autocomplete="off" autocorrect="off" autocapitalize="off" spellcheck="false" aria-label="Compiler options")
         .populararguments(title="Popular arguments")
           button.btn.btn-sm.btn-light.btn-outline-secondary.dropdown-toggle.dropdown-toggle-split.popular-arguments-btn(type="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false")
