+++ views/templates/panes/ir.pug
@@ -7,9 +7,8 @@ mixin optionButton(bind, isActive, text, title)
 #ir
   .top-bar.btn-toolbar.bg-light(role="toolbar")
     include ../../font-size
+    include ../widgets/compilation-options
     .btn-group.btn-group-sm.wrap(role="group")
-      button.btn.btn-sm.btn-light.input-group-text.prepend-options(data-trigger="click" style="cursor: pointer;" role="button" title="All compilation options")
-        span.status-icon
       .button-checkbox
         button.btn.btn-sm.btn-light.wrap-lines(type="button" title="Wrap lines" data-bind="wrap" aria-pressed="false" aria-label="Wrap lines")
           span Wrap lines
