+++ views/templates/widgets/compilation-options.pug
@@ -1,0 +1,2 @@
+button.btn.btn-sm.btn-light.input-group-text.prepend-options(data-trigger="click" style="cursor: pointer;" role="button" title="All compilation options")
+  span.status-icon
